requirements.txt
dgl==1.1.2+cu116
fvcore==0.1.5.post20221221
GitPython==3.1.40
GitPython==3.1.43
kernel==0.0.0
matplotlib==3.8.2
networkx==3.1
networkx==2.8.8
numpy==2.0.1
ogb==1.3.6
scikit_learn==1.3.0
scipy==1.14.0
six==1.16.0
torch==1.13.1
torch_geometric==2.5.3
torch_scatter==2.1.0
torch_sparse==0.6.15
torchdiffeq==0.2.3
torcheval==0.0.7
